Split Action types in separate file
A first step to implementing concrete actions
